refactor names to follow flutters naming convention
