+++ frontend/prototype/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:prototype/src/PostMortum/post_mortem_provider.dart';
+import 'package:prototype/src/post_mortem/post_mortem_provider.dart';
 import 'package:provider/provider.dart';
 
 import 'src/app.dart';
