+++ frontend/prototype/lib/src/unit_page/unit_page_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:prototype/src/device_settings/settings_view.dart';
-import 'package:prototype/src/PostMortum/PostMortem.dart';
+import 'package:prototype/src/post_mortem/post_mortem.dart';
 import 'package:prototype/src/Data/DataPage.dart';
 
 class UnitPageView extends StatelessWidget {
